src/astro_lab/astro_gnn/requirements.txt
# AstroGNN Requirements
# Kernabhängigkeiten für astronomische Punktwolken-Analyse

# PyTorch und Extensions
torch>=2.0.0
torch-geometric>=2.3.0
pytorch-lightning>=2.0.0
torchmetrics>=1.0.0

# Datenstrukturen
tensordict>=0.2.0
h5py>=3.8.0
pandas>=2.0.0
numpy>=1.24.0

# CLI und Utilities
click>=8.1.0
rich>=13.0.0
pyyaml>=6.0

# Visualisierung (optional)
matplotlib>=3.7.0
seaborn>=0.12.0

# Logging (optional)
wandb>=0.15.0
tensorboard>=2.13.0

# Wissenschaftliche Bibliotheken (optional)
astropy>=5.3.0  # Für astronomische Koordinaten-Transformationen
scikit-learn>=1.3.0  # Für zusätzliche Metriken
